Fixed lots of warnings mostly unused variables in function parameters
- one unused variable in a class MeshHierarchyGlobalMap can't be removed
  becasue it is used ???
- In CartesianHMesh comm is unused becasue the constructor uses the
  function argument not the member hence the name change to the
  parameter
- Had to cast lots of ints to size_t in asserts - hopefully won't casue
  any issues
- Had to add virtual destructors to some classes - I think so that a
  derived class accessed through a base pointer still calls the right
  one
- Reordering constructer lists to match the order of definition in the
  class

+++ include/neso_particles/cell_dat_compression.hpp
@@ -113,7 +113,7 @@ public:
     d_compress_layers_old.realloc_no_copy(npart);
     d_compress_layers_new.realloc_no_copy(npart);
 
-    NESOASSERT(this->d_npart_cell.size >= this->ncell,
+    NESOASSERT(this->d_npart_cell.size >= static_cast<std::size_t>(this->ncell),
                "Bad device_npart_cell length");
 
     const int ncell = this->ncell;
